Use SO_REUSEADDR on Linux and SO_REUSEPORT on macOS (#15)
... in team_message_receiver. At least on macOS it is the right thing
(as it would on other BSDs). I'm not sure if it is harmful to set
SO_REUSEADDR as well on macOS, and it probably doesn't hurt to use
SO_REUSEPORT on Linux, but it seems that SO_REUSEADDR on Linux is
already what SO_REUSEPORT is on BSD, just without TCP / load balancing
stuff that we don't want anyway.