Receive team messages on any address again
because IP doesn't work that way. Listening on the broadcast address of
the local interface doesn't work properly because the robots might send
on the limited broadcast address, which would not be received. Not sure
if it's worth to use SO_BINDTODEVICE.
(Partially revert 4b3e3b72c29360aa7ce29392ac546e42b0d0a806)