Implement fallback mode
What isn't correct yet is that the goalkeeper may be substituted against
a field player. Use at own risk.

The Switch Mode button is also in available in Champions Cup, and the
reason is simply that it requires a second row for the Motion in Standby
button so that the height does not switch when the area is rededicated
to the free kicks.