Add revamped version of TTS output
Thanking Arne Hasselbring for the idea of making a separate module
that communicates via tokio.
What I can do as a novice is still limited, but this should make it
easier to apply improvements later on.

This version uses a basic TTS library that interface with common
screen readers on multiple platforms (e.g. Speech Dispatcher).